+++ src/clear-cache.sh
@@ -1,14 +1,13 @@
-#!/usr/bin/env bash
+#!/usr/bin/env sh
 #
 # Frees up disk space by clearing caches of several package managers.
 
-# Exit immediately if a command exits or pipes a non-zero return code.
+# Exit immediately if a command exits with non-zero return code.
 #
 # Flags:
-#   -e: Exit immediately when a command pipeline fails.
-#   -o: Persist nonzero exit codes through a Bash pipe.
+#   -e: Exit immediately when a command fails.
 #   -u: Throw an error when an unset variable is encountered.
-set -eou pipefail
+set -eu
 
 #######################################
 # Show CLI help information.
@@ -20,16 +19,14 @@ usage() {
   case "${1}" in
     main)
       cat 1>&2 << EOF
-$(version)
 Frees up disk space by clearing caches of package managers.
 
-USAGE:
-    clear-cache [OPTIONS]
+Usage: clear-cache [OPTIONS]
 
-OPTIONS:
-        --debug      Show Bash debug traces
-    -h, --help       Print help information
-    -v, --version    Print version information
+Options:
+      --debug     Show shell debug traces
+  -h, --help      Print help information
+  -v, --version   Print version information
 EOF
       ;;
     *)
@@ -50,7 +47,7 @@ assert_cmd() {
   # Flags:
   #   -v: Only show file path of command.
   #   -x: Check if file exists and execute permission is granted.
-  if [[ ! -x "$(command -v "${1}")" ]]; then
+  if [ ! -x "$(command -v "${1}")" ]; then
     error "Cannot find required ${1} command on computer"
   fi
 }
@@ -61,7 +58,7 @@ assert_cmd() {
 #   Writes error message to stderr.
 #######################################
 error() {
-  local bold_red='\033[1;31m' default='\033[0m'
+  bold_red='\033[1;31m' default='\033[0m'
   printf "${bold_red}error${default}: %s\n" "${1}" >&2
   exit 1
 }
@@ -72,7 +69,7 @@ error() {
 #   Writes error message to stderr.
 #######################################
 error_usage() {
-  local bold_red='\033[1;31m' default='\033[0m'
+  bold_red='\033[1;31m' default='\033[0m'
   printf "${bold_red}error${default}: %s\n" "${1}" >&2
   printf "Run 'clear-cache --help' for usage.\n" >&2
   exit 2
@@ -82,57 +79,56 @@ error_usage() {
 # Clear cache of all package managers.
 #######################################
 clear_cache() {
-  local use_sudo
-
-  # Use sudo for system installation if user is not root.
-  if [[ "${EUID}" -ne 0 ]]; then
+  # Use sudo for system installation if user is not root. Do not use long form
+  # --user flag for id. It is not supported on MacOS.
+  if [ "$(id -u)" -ne 0 ]; then
     assert_cmd sudo
     use_sudo='true'
   fi
 
-  # Do not quote the sudo parameter expansion. Bash will error due to be being
+  # Do not quote the sudo parameter expansion. Script will error due to be being
   # unable to find the "" command.
-  if [[ -x "$(command -v apt-get)" ]]; then
+  if [ -x "$(command -v apt-get)" ]; then
     ${use_sudo:+sudo} apt-get clean --yes
   fi
 
-  if [[ -x "$(command -v brew)" ]]; then
+  if [ -x "$(command -v brew)" ]; then
     brew cleanup --prune all
   fi
 
-  if [[ -x "$(command -v dnf)" ]]; then
+  if [ -x "$(command -v dnf)" ]; then
     ${use_sudo:+sudo} dnf clean --assumeyes all
   fi
 
-  if [[ -x "$(command -v flatpak)" ]]; then
+  if [ -x "$(command -v flatpak)" ]; then
     ${use_sudo:+sudo} flatpak uninstall --assumeyes --unused
   fi
 
-  if [[ -x "$(command -v pacman)" ]]; then
+  if [ -x "$(command -v pacman)" ]; then
     ${use_sudo:+sudo} pacman --clean --sync
   fi
 
-  if [[ -x "$(command -v pkg)" ]]; then
+  if [ -x "$(command -v pkg)" ]; then
     ${use_sudo:+sudo} pkg clean --all --yes
   fi
 
-  if [[ -x "$(command -v zypper)" ]]; then
+  if [ -x "$(command -v zypper)" ]; then
     ${use_sudo:+sudo} zypper clean --all
   fi
 
-  if [[ -x "$(command -v docker)" ]]; then
+  if [ -x "$(command -v docker)" ]; then
     ${use_sudo:+sudo} docker system prune --force --volumes
   fi
 
-  if [[ -x "$(command -v npm)" ]]; then
+  if [ -x "$(command -v npm)" ]; then
     npm cache clean --force --loglevel error
   fi
 
-  if [[ -x "$(command -v nvm)" ]]; then
+  if [ -x "$(command -v nvm)" ]; then
     nvm cache clear
   fi
 
-  if [[ -x "$(command -v pip)" ]]; then
+  if [ -x "$(command -v pip)" ]; then
     pip cache purge
   fi
 }
@@ -143,7 +139,7 @@ clear_cache() {
 #   ClearCache version string.
 #######################################
 version() {
-  echo 'ClearCache 0.0.2'
+  echo 'ClearCache 0.1.0'
 }
 
 #######################################
@@ -151,7 +147,7 @@ version() {
 #######################################
 main() {
   # Parse command line arguments.
-  while [[ "$#" -gt 0 ]]; do
+  while [ "${#}" -gt 0 ]; do
     case "${1}" in
       --debug)
         set -o xtrace
@@ -172,7 +168,4 @@ main() {
   clear_cache
 }
 
-# Only run main if invoked as script. Otherwise import functions as library.
-if [[ "${BASH_SOURCE[0]}" == "$0" ]]; then
-  main "$@"
-fi
+main "$@"