+++ src/packup.ps1
@@ -2,14 +2,48 @@
 .SYNOPSIS
     Invokes upgrade commands to all installed package managers.
 #>
-[CmdletBinding()]
-Param()
 
 # Exit immediately if a PowerShell Cmdlet encounters an error.
 $ErrorActionPreference = 'Stop'
 
+# Show CLI help information.
+Function Usage() {
+    Write-Output @'
+Invokes upgrade commands to all installed package managers.
+
+Usage: packup [OPTIONS]
+
+Options:
+  -h, --help      Print help information
+  -v, --version   Print version information
+'@
+}
+
+# Print Packup version string.
+Function Version() {
+    Write-Output 'Packup 0.3.0'
+}
+
 # Script entrypoint.
 Function Main() {
+    $ArgIdx = 0
+
+    While ($ArgIdx -LT $Args[0].Count) {
+        Switch ($Args[0][$ArgIdx]) {
+            { $_ -In '-h', '--help' } {
+                Usage
+                Exit 0
+            }
+            { $_ -In '-v', '--version' } {
+                Version
+                Exit 0
+            }
+            Default {
+                $ArgIdx += 1
+            }
+        }
+    }
+
     If (Get-Command choco -ErrorAction SilentlyContinue) {
         choco upgrade --yes all
     }