+++ CHANGELOG.md
@@ -9,4 +9,4 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
-- Cloud Compute, Packup, TRsync, Tssh, and Virshx scripts.
+- ClearCache, Packup, Trsync, Tscp, and Tssh scripts.